+++ components/ToDo/List.jsx
@@ -1,11 +1,13 @@
+import { memo } from "react";
 import Item from "./Item";
 
 const List = (props) => {
   const { toDos } = props;
+  console.log("render List");
 
   return toDos.map((toDo) => (
     <Item key={toDo.id} id={toDo.id} toDo={toDo} {...props} />
   ));
 };
 
-export default List;
+export default memo(List);
